backend/requirements.txt
# FastAPI and related
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Redis for caching
redis==5.0.1
aioredis==2.0.1

# Celery for background tasks
celery==5.3.4
flower==2.0.1

# HTTP requests
httpx==0.25.2
aiohttp==3.9.1

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Data processing
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2

# Geospatial data
geopandas==0.14.1
shapely==2.0.2
folium==0.15.0

# Satellite imagery
rasterio==1.3.9
sentinelhub==3.9.2

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1